requirements.txt
tensorflow-cpu==2.9.0
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
numpy==1.19.5
matplotlib==3.3.4
python-dateutil==2.8.2
pytest==7.0.1
pycodestyle==2.8.0
pep8==1.7.1
scikit-image==0.19.3
scipy==1.7.3
jupyter
notebook